USB driver modified to set a flag when a USB connection is detected. When the board is plugged in via USB it will not enter power saving mode. Removed unused shift_disable function from shift register class. Added sleep functionality to main - sleep is entered in mode 0 or 4 if speed drops below approx 60 rpm Improved power usage by disabling unused modules on the AVR (timer2, ADC, SPI, etc) Modified ISRs for the display and the hall effect.

+++ Firmware/HallEffect.h
@@ -7,6 +7,7 @@
 #include <avr/interrupt.h>
 
 #include "globalDefs.h"
+#include "display.h"
 
 /* Set up the ports/pin for the hall effect switch. */
 #define HALL_REG PORTD
@@ -42,6 +43,6 @@ uint32_t get_cycles(void);
  * Get the distance travelled since the board was powered on.
  * Returns: distance travelled (in metres).
  */
-uint16_t get_distance(void);
+uint32_t get_distance(void);
 
 #endif
